added file containing crime data for US
This .csv file is in the same format as the NJ and CA crime data, and from the FBI UCR website, but rather than containing data for only one state, it contains data for all 50 states as well as the District of Columbia. I also added a column called "State" with state names that signify which state each case in the data belongs to.