final project .Rmd file
this is for the last stage of the final project, and I am almost finished with it. after several more commits, I should be ready for the final submission and be able to create a guide sheet for grading. some changes that come with this final step are removing the NJ and CA crime data sets and introducing an all-encompassing "USCrime" data table that contains crime data for all 50 states as well as the District of Columbia.